requirements.txt
torch==2.8.0
torchrl==0.9.2
tensordict==0.9.1
ruamel.yaml==0.17.4
moviepy==1.0.3
einops==0.3.0
protobuf==3.20.0
mujoco==3.3.0
dm_control==1.0.28
metaworld==3.0.0
memory_maze==1.0.3
crafter==1.8.0
opencv-python==4.7.0.72
numpy==1.26.0
hydra-core==1.3.2
ale_py==0.11.2
gymnasium==1.3.0
popgym==1.0.6
bsuite==0.3.5; python_version < "3.12"
pre-commit==4.3.0
gym==0.26.2 # required for memory_maze
wandb==0.19.1
matplotlib==3.10.8
# Optional: enables the Triton SSD fast path for the mamba2 backbone on CUDA hosts.
# The pure-PyTorch fallback in rssm_mamba2.py is used when this isn't installed.
mamba-ssm>=2.2.2 ; sys_platform == "linux" and platform_machine == "x86_64"
